chore: temporary structure with incomplete caputures
The first problem, if we put a restriction on nested rules, we get the problem that there can be several roots, and they can be all of different types

If for example we loaded bytes for validations, but made root rules for both bytes and strings, should we allow auto-conversion to strings (without UTF8 validation). If you give such permission, then it is strange why you should use a restriction on a certain type of rules

Alternatively, you can put Rule as a new structure RuleBytes, it's easier to control types, or specify root checking in TemplateValidator, which will check that all roots are of the same type.